requirements.txt
Pillow==9.1.1
numpy==1.22.4
face_recognition==1.3.0
streamlit==1.9.2